src/main/java/com/yevheniiStasiuk/mvp/service/GameReaderService.java
package com.yevheniiStasiuk.mvp.service;

import org.springframework.stereotype.Service;

import java.io.IOException;
import java.nio.charset.StandardCharsets;
import java.nio.file.Files;
import java.nio.file.Path;
import java.nio.file.Paths;
import java.util.List;

@Service
public class GameReaderService {
    public List<String> readGame(String fileName) throws IOException {
        final Path path = Paths.get(fileName);
        return Files.readAllLines(path, StandardCharsets.UTF_8);
    }
}
